requirements.txt
appnope==0.1.0
argon2-cffi==20.1.0
asgiref==3.2.10
attrs==19.3.0
backcall==0.2.0
beautifulsoup4==4.9.1
bleach==3.1.5
boto3==1.14.43
botocore==1.17.43
certifi==2020.6.20
cffi==1.14.1
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
decorator==4.4.2
defusedxml==0.6.0
Django==3.1
django-crontab==0.7.1
django-debug-toolbar==3.1.1
django-debug-toolbar-request-history==0.1.3
django-environ==0.4.5
django-extensions==3.0.5
django-filter==2.3.0
django-inmemorystorage==0.1.1
django-redis==4.12.1
django-storages==1.9.1
djangorestframework==3.11.1
docutils==0.15.2
drf-action-serializer==1.0.5
drf-nested-routers==0.91
drf-yasg==1.17.1
drf-yasg-examples==0.1.1
entrypoints==0.3
gunicorn==20.0.4
idna==2.10
inflection==0.5.1
ipykernel==5.3.4
ipython==7.17.0
ipython-genutils==0.2.0
itypes==1.2.0
jedi==0.17.2
Jinja2==2.11.2
jmespath==0.10.0
jsonschema==3.2.0
jupyter-client==6.1.6
jupyter-core==4.6.3
MarkupSafe==1.1.1
mistune==0.8.4
model-bakery==1.1.1
munch==2.5.0
nbconvert==5.6.1
nbformat==5.0.7
notebook==6.1.3
packaging==20.4
pandocfilters==1.4.2
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
prometheus-client==0.8.0
prompt-toolkit==3.0.6
psycopg2-binary==2.8.5
ptyprocess==0.6.0
pycparser==2.20
Pygments==2.6.1
pyparsing==2.4.7
pyrsistent==0.16.0
python-dateutil==2.8.1
pytz==2020.1
pyzmq==19.0.2
redis==3.5.3
requests==2.24.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
s3transfer==0.3.3
selenium==3.141.0
Send2Trash==1.5.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
terminado==0.8.3
testpath==0.4.4
tornado==6.0.4
traitlets==4.3.3
uritemplate==3.0.1
urllib3==1.25.10
wcwidth==0.2.5
webencodings==0.5.1
